add a Registry trait
